Change num to int

+++ src/main/valuepair.c
@@ -934,7 +934,7 @@ int radius_map2request(REQUEST *request, value_pair_map_t const *map,
 		       UNUSED char const *src, radius_tmpl_getvalue_t func, void *ctx)
 {
 	int rcode;
-	int16_t num;
+	int num;
 	VALUE_PAIR **list, *vp, *dst, *head = NULL;
 	bool found = false;
 	REQUEST *context;
@@ -1670,7 +1670,7 @@ VALUE_PAIR *radius_vpt_get_vp(REQUEST *request, value_pair_tmpl_t const *vpt)
 	 */
 	case VPT_TYPE_ATTR:
 	{
-		int16_t num;
+		int num;
 		vp_cursor_t cursor;
 
 		if (vpt->vpt_num == NUM_ANY) {
