+++ src/main/xlat.c
@@ -61,7 +61,7 @@ struct xlat_exp {
 
 	DICT_ATTR const *da;	//!< the name of the dictionary attribute
 
-	int16_t num;		//!< attribute number
+	int num;		//!< attribute number
 	int8_t tag;		//!< attribute tag
 	pair_lists_t list;	//!< list of which attribute
 	request_refs_t ref;	//!< outer / this / ...
@@ -1568,7 +1568,7 @@ static ssize_t xlat_tokenize_request(REQUEST *request, char const *fmt, xlat_exp
 
 
 static char *xlat_getvp(TALLOC_CTX *ctx, REQUEST *request, pair_lists_t list, DICT_ATTR const *da,
-			int8_t tag, int16_t num, bool return_null)
+			int8_t tag, int num, bool return_null)
 {
 	VALUE_PAIR *vp, *vps = NULL, *myvp = NULL;
 	RADIUS_PACKET *packet = NULL;
